feat: enhance login handler with improved error handling and secure session management

+++ backend/handler/loginhandler.go
@@ -5,11 +5,13 @@ import (
 	"errors"
 	"log"
 	"net/http"
+	"strconv"
 	"sync"
 	"time"
 
 	"github.com/jesee-kuya/forum/backend/models"
 	"github.com/jesee-kuya/forum/backend/repositories"
+	"github.com/jesee-kuya/forum/backend/util"
 	"golang.org/x/crypto/bcrypt"
 )
 
@@ -28,23 +30,24 @@ func LoginHandler(w http.ResponseWriter, r *http.Request) {
 
 	err := json.NewDecoder(r.Body).Decode(&data)
 	if err != nil {
-		w.Header().Set("Content-Type", "application/json")
-		w.WriteHeader(http.StatusInternalServerError)
-		json.NewEncoder(w).Encode(map[string]string{
-			"error": "unknown error occured. Try again later",
-		})
+		util.ErrorHandler(w, "Invalid Request", http.StatusBadRequest)
 		return
 	}
 
 	user, session, err := Login(data.Password, data.Email)
 	if err != nil {
-		w.Header().Set("Content-Type", "application/json")
-		w.WriteHeader(http.StatusInternalServerError)
-		json.NewEncoder(w).Encode(map[string]string{
-			"error": "unknown error occured. Try again later",
-		})
+		util.ErrorHandler(w, "Invalid Credentials", http.StatusUnauthorized)
 		return
 	}
+
+	// set a secure session cookie
+	http.SetCookie(w, &http.Cookie{
+		Name:     "session_token",
+		Value:    session,
+		Expires:  time.Now().Add(24 * time.Hour),
+		Path:     "/",
+		HttpOnly: true,
+	})
 
 	log.Println(SessionStore)
 
@@ -85,7 +88,19 @@ func Login(password, email string) (models.User, string, error) {
 	}
 
 	if user.ID != 0 {
-		DeleteSession(user.ID)
+		mu.Lock()
+		userIDStr := strconv.Itoa(user.ID)
+		_, exists := SessionStore[userIDStr]
+		mu.Unlock()
+
+		if exists {
+			DeleteSession(user.ID)
+			err = repositories.DeleteSessionByUser(user.ID)
+			if err != nil {
+				log.Printf("Failed to delete session token: %v", err)
+				return user, "", errors.New("an Unexpected Error Occurred. Try Again Later")
+			}
+		}
 	}
 
 	err = repositories.DeleteSessionByUser(user.ID)
